HACK: Bypass hashing with only one destination
This gives a really nice speedup when receiving a constant stream of
metrics (>10k/sec).

It is not scalable however, maybe we could store the result of the
hashing in a cache.

Here's a cProfile sample of the affected calls, these calls are avoided with this patch.
ncalls  tottime  percall  cumtime  percall filename:lineno(function)
9973508   17.726    0.000   41.087    0.000 /opt/graphite/lib/carbon/hashing.py:41(get_nodes)
4986854    8.300    0.000   16.505    0.000 /opt/graphite/lib/carbon/hashing.py:16(compute_ring_position)
4986855    4.743    0.000    4.743    0.000 {_hashlib.openssl_md5}
4986754    3.680    0.000    3.680    0.000 {_bisect.bisect_left}
4986854    3.461    0.000    3.461    0.000 {method 'hexdigest' of '_hashlib.HASH' objects}